requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator==2.1.0.post1
PyMySQL==1.1.0
cryptography==42.0.2